+++ modules/matterhorn-composer-service-api/src/main/java/org/opencastproject/composer/api/ComposerService.java
@@ -140,7 +140,7 @@ public interface ComposerService {
    * @throws MediaPackageException
    *           if the mediapackage is invalid
    */
-  Job trim(Track sourceTrack, String profileId, double start, double duration) throws EncoderException,
+  Job trim(Track sourceTrack, String profileId, long start, long duration) throws EncoderException,
           MediaPackageException;
 
   /**
